backend/pyproject.toml
[project]
name = "fastapi-sqlserver"
version = "0.1.0"
description = ""
authors = [
    {name = "dapkou",email = "dapkou@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "pyodbc (>=5.3.0,<6.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "bcrypt (<4)"
]

[tool.poetry]
packages = [{include = "accounts", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"